+++ consensus/helpers.go
@@ -44,18 +44,18 @@ func (m *consensusModule) getQuorumCertificate(height uint64, step typesCons.Hot
 			return nil, err
 		}
 		if msg.GetPartialSignature() == nil {
-			m.logger.Warn().Fields(msgToLoggingFields(msg)).Msg("No partial signature found which should not happen...")
+			m.logger.Warn().Fields(hotstuffMsgToLoggingFields(msg)).Msg("No partial signature found which should not happen...")
 			continue
 		}
 		if msg.GetHeight() != height || msg.GetStep() != step || msg.GetRound() != round {
-			m.logger.Warn().Fields(msgToLoggingFields(msg)).Msg("Message in pool does not match (height, step, round) of QC being generated")
+			m.logger.Warn().Fields(hotstuffMsgToLoggingFields(msg)).Msg("Message in pool does not match (height, step, round) of QC being generated")
 			continue
 		}
 
 		ps := msg.GetPartialSignature()
 		if ps.Signature == nil || ps.Address == "" {
 
-			m.logger.Warn().Fields(msgToLoggingFields(msg)).Msg("Partial signature is incomplete which should not happen...")
+			m.logger.Warn().Fields(hotstuffMsgToLoggingFields(msg)).Msg("Partial signature is incomplete which should not happen...")
 			continue
 		}
 		pss = append(pss, msg.GetPartialSignature())
@@ -146,7 +146,7 @@ func protoHash(m proto.Message) string {
 func (m *consensusModule) sendToLeader(msg *typesCons.HotstuffMessage) {
 	leaderId := m.leaderId
 
-	loggingFields := msgToLoggingFields(msg)
+	loggingFields := hotstuffMsgToLoggingFields(msg)
 	loggingFields["src"] = m.nodeId
 	loggingFields["dst"] = leaderId
 	m.logger.Debug().Fields(loggingFields).Msg("✉️ About to try sending hotstuff message ✉️")
@@ -179,7 +179,7 @@ func (m *consensusModule) sendToLeader(msg *typesCons.HotstuffMessage) {
 // Star-like (O(n)) broadcast - send to all nodes directly
 // INVESTIGATE: Re-evaluate if we should be using our structured broadcast (RainTree O(log3(n))) algorithm instead
 func (m *consensusModule) broadcastToValidators(msg *typesCons.HotstuffMessage) {
-	m.logger.Info().Fields(msgToLoggingFields(msg)).Msg("📣 Broadcasting message 📣")
+	m.logger.Info().Fields(hotstuffMsgToLoggingFields(msg)).Msg("📣 Broadcasting message 📣")
 
 	anyConsensusMessage, err := codec.GetCodec().ToAny(msg)
 	if err != nil {
@@ -219,7 +219,7 @@ func (m *consensusModule) isReplica() bool {
 }
 
 func (m *consensusModule) electNextLeader(msg *typesCons.HotstuffMessage) error {
-	loggingFields := msgToLoggingFields(msg)
+	loggingFields := hotstuffMsgToLoggingFields(msg)
 	m.logger.Info().Fields(loggingFields).Msg("About to elect the next leader")
 
 	m.leaderId = nil
@@ -271,14 +271,6 @@ func (m *consensusModule) getValidatorsAtHeight(height uint64) ([]*coreTypes.Act
 	return readCtx.GetAllValidators(int64(height))
 }
 
-func msgToLoggingFields(msg *typesCons.HotstuffMessage) map[string]any {
-	return map[string]any{
-		"height": msg.GetHeight(),
-		"round":  msg.GetRound(),
-		"step":   typesCons.StepToString[msg.GetStep()],
-	}
-}
-
 // TODO: This is a temporary solution, cache this in Consensus module. This field will be populated once with a single query to the persistence module.
 func (m *consensusModule) IsValidator() (bool, error) {
 	validators, err := m.getValidatorsAtHeight(m.CurrentHeight())
@@ -293,4 +285,12 @@ func (m *consensusModule) IsValidator() (bool, error) {
 	}
 
 	return false, nil
+}
+
+func hotstuffMsgToLoggingFields(msg *typesCons.HotstuffMessage) map[string]any {
+	return map[string]any{
+		"height": msg.GetHeight(),
+		"round":  msg.GetRound(),
+		"step":   typesCons.StepToString[msg.GetStep()],
+	}
 }