+++ shared/CHANGELOG.md
@@ -7,9 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.50] - 2023-04-12
+## [0.0.0.51] - 2023-04-12
 
 - Removed *temporary* `shared/p2p` package; consolidated into `p2p`
+
+## [0.0.0.50] - 2023-04-10
+
+- Added `modules.ModuleFactoryWithOptions` interface
+- Added factory interfaces:
+    - `modules.FactoryWithRequired`
+    - `modules.FactoryWithOptions`
+    - `modules.FactoryWithRequiredAndOptions`
+- Embedded `ModuleFactoryWithOptions` in `Module` interface
+- Switched mock generation to use reflect mode for effected interfaces (embedders)
 
 ## [0.0.0.49] - 2023-04-07